move subcomponents to own files
