+++ packages/components/src/partials/BaseDropdown/components/ItemsContainer/index.ts
@@ -1,0 +1,2 @@
+export * from './ItemsContainer';
+export * from './ItemsContainer.types';
