+++ packages/components/src/partials/BaseDropdown/components/Search/Search.types.ts
@@ -1,0 +1,6 @@
+export interface SearchProps {
+    isVisible: boolean;
+    setSearchTerm: (value: string) => void;
+    searchRef: React.RefObject<HTMLInputElement | null>;
+    searchTerm: string;
+}
