+++ packages/components/src/partials/BaseDropdown/components/Search/index.ts
@@ -1,0 +1,2 @@
+export * from './Search';
+export * from './Search.types';
